adjusted acceptance test to test startup failure with unavailable SC queue
