+++ src/ServiceControl.Plugin.Nsb5.Heartbeat.AcceptanceTests/ServiceControl.Plugin.Nsb5.Heartbeat.AcceptanceTests.csproj
@@ -45,6 +45,7 @@
     <Reference Include="System" />
     <Reference Include="System.Configuration" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Messaging" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -54,7 +55,7 @@
   <ItemGroup>
     <Compile Include="DefaultServer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="When_using_a_custom_circuit_breaker_timeout.cs" />
+    <Compile Include="When_service_control_queue_unavailable_at_startup.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
